refactor: 💡 Creates protocols sharing code in dialog builders
