+++ Sources/Core/ViewStyling/Dialog/NatDialogController+ButtonConfiguration.swift
@@ -1,6 +1,7 @@
-extension NatDialogController {
+extension NatDialogController { //TODO REVIEW
     struct ButtonConfiguration {
         let title: String
+        let style: NatButton.Style
         let action: () -> Void
     }
 }
