+++ Sources/Public/Components/Dialog/Builders/NatDialogController+AlertStyleBuilder.swift
@@ -1,69 +1,38 @@
 extension NatDialogController {
     public final class AlertStyleBuilder {
 
-        // MARK: - Private properties
+        // MARK: - DialogBuilder
 
-        private var titleView: UIView?
-        private var bodyView: UIView?
-        private var primaryButtonConfiguration: ButtonConfiguration?
-        private var secondaryButtonConfiguration: ButtonConfiguration?
+        public let viewModel = ViewModel()
 
         // MARK: - Public methods
 
-//        public func configure(title: String) -> Self {
-//            titleView = DialogStyle.createLabelForTitle(title: title)
-//
-//            return self
-//        }
-
-        public func configure(body: String) -> Self {
-            bodyView = DialogStyle.createLabelForBody(body: body)
-
-            return self
-        }
-
-        public func configure(body: UIView) -> Self {
-            bodyView = body
-
-            return self
-        }
-
         public func configure(primaryButtonTitle title: String, withAction action: @escaping () -> Void) -> Self {
-            primaryButtonConfiguration = .init(title: title, action: action)
+            let footerView = DialogStyle.FooterView()
+
+            footerView.configure(primaryButton: .init(
+                title: title,
+                style: .text,
+                action: action)
+            )
+
+            viewModel.footerView = footerView
 
             return self
         }
 
         public func configure(secondaryButtonTitle title: String, withAction action: @escaping () -> Void) -> Self {
-            secondaryButtonConfiguration = .init(title: title, action: action)
+            viewModel.footerView?.configure(secondaryButton: .init(
+                title: title,
+                style: .text,
+                action: action)
+            )
 
             return self
         }
-
-        public func build() -> NatDialogController {
-            var views: [UIView] = []
-
-            if let view = titleView {
-
-                views.append(view)
-            }
-
-            if let view = bodyView {
-                views.append(view)
-            }
-
-            if let primaryAction = primaryButtonConfiguration {
-                let footerView = DialogAlertStyle.FooterView()
-                views.append(footerView)
-
-                footerView.configure(primaryButton: primaryAction)
-
-                if let secondaryAction = secondaryButtonConfiguration {
-                    footerView.configure(secondaryButton: secondaryAction)
-                }
-            }
-
-            return .init(views: views)
-        }
     }
 }
+
+extension NatDialogController.AlertStyleBuilder: DialogBuilder {}
+extension NatDialogController.AlertStyleBuilder: DialogTitleConfigurator {}
+extension NatDialogController.AlertStyleBuilder: DialogBodyConfigurator {}
