+++ Sources/Public/Components/Dialog/Builders/Protocols/DialogBuilder.swift
@@ -1,0 +1,9 @@
+public protocol DialogBuilder {
+    var viewModel: NatDialogController.ViewModel { get }
+}
+
+extension DialogBuilder {
+    public func build() -> NatDialogController {
+        return .init(viewModel: viewModel)
+    }
+}
