+++ Sources/Public/Components/Dialog/Builders/Protocols/DialogTitleConfigurator.swift
@@ -1,0 +1,9 @@
+public protocol DialogTitleConfigurator: DialogBuilder {}
+
+extension DialogTitleConfigurator {
+    public func configure(title: String) -> Self {
+        viewModel.titleView = DialogStyle.createLabelForTitle(title: title)
+
+        return self
+    }
+}
