+++ Sources/Public/Components/Dialog/DialogViewModel.swift
@@ -1,0 +1,8 @@
+extension NatDialogController {
+    public class ViewModel {
+        var titleView: UIView?
+        var bodyView: UIView?
+        var footerView: DialogStyle.FooterView?
+    }
+
+}
